arch/tools/pyproject.toml
[tool.poetry]
name = "archgw"
version = "0.3.3"
description = "Python-based CLI tool to manage Arch Gateway."
authors = ["Katanemo Labs, Inc."]
packages = [
    { include = "cli" }
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
archgw_modelserver = "^0.3.3"
click = "^8.1.7"
jinja2 = "^3.1.4"
jsonschema = "^4.23.0"
setuptools = "75.5.0"
pyyaml = "^6.0.2"

[tool.poetry.scripts]
archgw = "cli.main:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = ["-v"]
